:sparkles: Add UI components
Add Navbar component, Header component, Footer component, and ScrollToTop component. Add stylesheet for navigation menu links.
