+++ src/styles/components/navlinks.css
@@ -1,0 +1,17 @@
+a.middle-underline {
+    @apply relative;
+}
+
+a.middle-underline:after {
+    @apply absolute w-full bottom-0 right-0 bg-gradient-to-r from-cyan-400 to-lightblue-600 invisible;
+    content: "";
+    height: 3px;
+    transform: scaleX(0);
+    transition: all 0.325s ease-in-out;
+    border-radius: 0 100px 0 100px;
+}
+
+a.middle-underline:hover:after {
+    @apply visible;
+    transform: scaleX(1);
+}
